@@ -22,11 +22,10 @@ export 'models/referral_choices.dart';
 export 'models/space_schema.dart';
 export 'models/input.dart';
 export 'models/meeting_provider_enum.dart';
+export 'models/mobile_space_detail_schema.dart';
 export 'models/next_event_schema.dart';
-export 'models/paged_space_detail_schema.dart';
-export 'models/space_detail_schema.dart';
+export 'models/paged_mobile_space_detail_schema.dart';
 export 'models/event_detail_schema.dart';
-export 'models/event_space_schema.dart';
 export 'models/summary_spaces_schema.dart';
 export 'models/blog_post_list_schema.dart';
 export 'models/paged_blog_post_list_schema.dart';
