@@ -10,14 +10,12 @@ EventDetailSchema _$EventDetailSchemaFromJson(Map<String, dynamic> json) =>
     EventDetailSchema(
       slug: json['slug'] as String,
       title: json['title'] as String,
-      space: EventSpaceSchema.fromJson(json['space'] as Map<String, dynamic>),
-      spaceTitle: json['space_title'] as String,
-      description: json['description'] as String,
-      price: (json['price'] as num).toInt(),
+      space: MobileSpaceDetailSchema.fromJson(
+        json['space'] as Map<String, dynamic>,
+      ),
+      content: json['content'] as String,
       seatsLeft: (json['seats_left'] as num).toInt(),
       duration: (json['duration'] as num).toInt(),
-      recurring: json['recurring'] as String,
-      subscribers: (json['subscribers'] as num).toInt(),
       start: DateTime.parse(json['start'] as String),
       attending: json['attending'] as bool,
       open: json['open'] as bool,
@@ -42,13 +40,9 @@ Map<String, dynamic> _$EventDetailSchemaToJson(
   'slug': instance.slug,
   'title': instance.title,
   'space': instance.space,
-  'space_title': instance.spaceTitle,
-  'description': instance.description,
-  'price': instance.price,
+  'content': instance.content,
   'seats_left': instance.seatsLeft,
   'duration': instance.duration,
-  'recurring': instance.recurring,
-  'subscribers': instance.subscribers,
   'start': instance.start.toIso8601String(),
   'attending': instance.attending,
   'open': instance.open,
