@@ -3,7 +3,7 @@ import 'dart:convert';
 import 'package:flutter/foundation.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:totem_app/api/models/event_detail_schema.dart';
-import 'package:totem_app/api/models/space_detail_schema.dart';
+import 'package:totem_app/api/models/mobile_space_detail_schema.dart';
 import 'package:totem_app/api/models/space_schema.dart';
 import 'package:totem_app/api/models/summary_spaces_schema.dart';
 import 'package:totem_app/core/config/consts.dart';
@@ -64,7 +64,7 @@ class CacheService {
 
   // Spaces cache methods
 
-  Future<void> saveSpaces(List<SpaceDetailSchema> spaces) {
+  Future<void> saveSpaces(List<MobileSpaceDetailSchema> spaces) {
     return write(
       key: AppConsts.storageSpacesListKey,
       value: {
@@ -73,11 +73,12 @@ class CacheService {
     );
   }
 
-  Future<List<SpaceDetailSchema>?> getSpaces() async {
+  Future<List<MobileSpaceDetailSchema>?> getSpaces() async {
     final data = await read(AppConsts.storageSpacesListKey);
     return (data?['spaces'] as List?)
-        ?.map<SpaceDetailSchema>(
-          (json) => SpaceDetailSchema.fromJson(json as Map<String, dynamic>),
+        ?.map<MobileSpaceDetailSchema>(
+          (json) =>
+              MobileSpaceDetailSchema.fromJson(json as Map<String, dynamic>),
         )
         .toList();
   }
