@@ -5,10 +5,9 @@ import 'package:flutter_html/flutter_html.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:go_router/go_router.dart';
 import 'package:intl/intl.dart';
-import 'package:markdown/markdown.dart' as markdown;
 import 'package:share_plus/share_plus.dart';
 import 'package:totem_app/api/models/event_detail_schema.dart';
-import 'package:totem_app/api/models/space_detail_schema.dart';
+import 'package:totem_app/api/models/mobile_space_detail_schema.dart';
 import 'package:totem_app/core/config/app_config.dart';
 import 'package:totem_app/core/config/theme.dart';
 import 'package:totem_app/core/services/analytics_service.dart';
@@ -63,7 +62,7 @@ class _SpaceDetailScreenState extends ConsumerState<SpaceDetailScreen> {
     final String? effectiveEventSlug =
         widget.eventSlug ??
         spaceAsync.maybeWhen(
-          data: (space) => space.nextEvent?.slug,
+          data: (space) => space.nextEvents.firstOrNull?.slug,
           orElse: () => null,
         );
 
@@ -268,7 +267,9 @@ class _SpaceDetailScreenState extends ConsumerState<SpaceDetailScreen> {
                           Html(
                             data: space.shortDescription.trim().isNotEmpty
                                 ? space.shortDescription
-                                : space.content,
+                                : space.content.trim().length > 200
+                                ? '${space.content.substring(0, 200)}...'
+                                : space.content.trim(),
                             style: {
                               ...AppTheme.htmlStyle,
                               'body': Style(
@@ -389,7 +390,7 @@ class _SpaceDetailScreenState extends ConsumerState<SpaceDetailScreen> {
 
   Future<void> _showAboutSpaceSheet(
     BuildContext context,
-    SpaceDetailSchema space,
+    MobileSpaceDetailSchema space,
   ) {
     return showModalBottomSheet(
       context: context,
@@ -404,7 +405,7 @@ class _SpaceDetailScreenState extends ConsumerState<SpaceDetailScreen> {
 
   Future<void> _showSessionSheet(
     BuildContext context,
-    SpaceDetailSchema space,
+    MobileSpaceDetailSchema space,
     EventDetailSchema event,
   ) {
     return showModalBottomSheet(
@@ -422,7 +423,7 @@ class _SpaceDetailScreenState extends ConsumerState<SpaceDetailScreen> {
 class AboutSpaceSheet extends StatelessWidget {
   const AboutSpaceSheet({required this.space, super.key});
 
-  final SpaceDetailSchema space;
+  final MobileSpaceDetailSchema space;
 
   @override
   Widget build(BuildContext context) {
@@ -517,7 +518,9 @@ class AboutSpaceSheet extends StatelessWidget {
                     ],
                   ),
                   Html(
-                    data: markdown.markdownToHtml(space.content),
+                    data: space.content,
+                    style: {...AppTheme.compactHtmlStyle},
+                    extensions: [TotemImageHtmlExtension()],
                     shrinkWrap: true,
                     onLinkTap: (url, _, _) async {
                       if (url != null) {
@@ -543,8 +546,6 @@ class AboutSpaceSheet extends StatelessWidget {
                         }
                       }
                     },
-                    style: {...AppTheme.compactHtmlStyle},
-                    extensions: [TotemImageHtmlExtension()],
                   ),
                 ],
               ),
@@ -559,7 +560,7 @@ class AboutSpaceSheet extends StatelessWidget {
 class SessionSheet extends StatelessWidget {
   const SessionSheet({required this.space, required this.event, super.key});
 
-  final SpaceDetailSchema space;
+  final MobileSpaceDetailSchema space;
   final EventDetailSchema event;
 
   @override
@@ -638,7 +639,7 @@ class SessionSheet extends StatelessWidget {
                   ],
                 ),
                 Html(
-                  data: space.content,
+                  data: event.content,
                   shrinkWrap: true,
                   style: AppTheme.compactHtmlStyle,
                   extensions: [TotemImageHtmlExtension()],
