@@ -1,14 +1,13 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:sliver_tools/sliver_tools.dart';
-import 'package:totem_app/api/models/space_detail_schema.dart';
+import 'package:totem_app/api/models/mobile_space_detail_schema.dart';
 import 'package:totem_app/features/spaces/repositories/space_repository.dart';
 import 'package:totem_app/features/spaces/widgets/filter.dart';
 import 'package:totem_app/features/spaces/widgets/space_card.dart';
 import 'package:totem_app/shared/totem_icons.dart';
 import 'package:totem_app/shared/widgets/empty_indicator.dart';
 import 'package:totem_app/shared/widgets/error_screen.dart';
-import 'package:totem_app/shared/widgets/loading_indicator.dart';
 import 'package:totem_app/shared/widgets/totem_icon.dart';
 
 // Provider to track the selected category filter
@@ -131,12 +130,20 @@ class SpacesDiscoveryScreen extends ConsumerWidget {
           showHomeButton: false,
           onRetry: () => ref.refresh(listSpacesProvider.future),
         ),
-        loading: () => const LoadingIndicator(),
+        loading: () {
+          return ListView.separated(
+            physics: const NeverScrollableScrollPhysics(),
+            padding: const EdgeInsetsDirectional.all(16),
+            itemBuilder: (_, _) => SpaceCard.shimmer(),
+            separatorBuilder: (_, _) => const SizedBox(height: 16),
+            itemCount: (MediaQuery.heightOf(context) / 100).round(),
+          );
+        },
       ),
     );
   }
 
-  List<String> _extractCategories(List<SpaceDetailSchema> spaces) {
+  List<String> _extractCategories(List<MobileSpaceDetailSchema> spaces) {
     final categories =
         spaces
             .map((space) => space.category)
