@@ -2,9 +2,13 @@ import 'package:auto_size_text/auto_size_text.dart';
 import 'package:cached_network_image/cached_network_image.dart';
 import 'package:flutter/material.dart';
 import 'package:go_router/go_router.dart';
+import 'package:shimmer/shimmer.dart';
 import 'package:totem_app/api/models/event_detail_schema.dart';
+import 'package:totem_app/api/models/meeting_provider_enum.dart';
+import 'package:totem_app/api/models/mobile_space_detail_schema.dart';
 import 'package:totem_app/api/models/next_event_schema.dart';
-import 'package:totem_app/api/models/space_detail_schema.dart';
+import 'package:totem_app/api/models/profile_avatar_type_enum.dart';
+import 'package:totem_app/api/models/public_user_schema.dart';
 import 'package:totem_app/navigation/route_names.dart';
 import 'package:totem_app/shared/assets.dart';
 import 'package:totem_app/shared/date.dart';
@@ -15,32 +19,75 @@ import 'package:totem_app/shared/utils.dart';
 import 'package:totem_app/shared/widgets/space_gradient_mask.dart';
 import 'package:totem_app/shared/widgets/user_avatar.dart';
 
-SpaceDetailSchema _spaceDetailFromEventDetailSchema(EventDetailSchema event) {
-  return SpaceDetailSchema(
-    slug: event.space.slug!,
+MobileSpaceDetailSchema _spaceDetailFromEventDetailSchema(
+  EventDetailSchema event,
+) {
+  return MobileSpaceDetailSchema(
+    slug: event.space.slug,
     title: event.space.title,
-    imageLink: event.space.image,
+    imageLink: event.space.imageLink,
     content: event.space.content,
-    shortDescription: event.space.shortDescription ?? '',
+    shortDescription: event.space.shortDescription,
     author: event.space.author,
     recurring: event.space.recurring,
-    price: event.price,
-    subscribers: event.subscribers,
-    nextEvent: NextEventSchema(
-      start: event.start,
-      link: event.calLink,
-      seatsLeft: event.seatsLeft,
-      slug: event.slug,
-      title: event.title,
-      attending: event.attending,
-      calLink: event.calLink,
-      cancelled: event.cancelled,
-      duration: event.duration,
-      joinable: event.joinable,
-      meetingProvider: event.meetingProvider,
-      open: event.open,
+    price: event.space.price,
+    subscribers: event.space.subscribers,
+    nextEvents: [
+      NextEventSchema(
+        start: event.start,
+        link: event.calLink,
+        seatsLeft: event.seatsLeft,
+        slug: event.slug,
+        title: event.title,
+        attending: event.attending,
+        calLink: event.calLink,
+        cancelled: event.cancelled,
+        duration: event.duration,
+        joinable: event.joinable,
+        meetingProvider: event.meetingProvider,
+        open: event.open,
+      ),
+    ],
+    category: event.space.category,
+  );
+}
+
+MobileSpaceDetailSchema _dummySpaceDetail() {
+  return MobileSpaceDetailSchema(
+    slug: 'dummy-space',
+    title: 'Dummy Space',
+    imageLink: 'https://placehold.co/400',
+    shortDescription: 'Dummy Space',
+    content: 'Dummy Space',
+    author: PublicUserSchema(
+      profileAvatarType: ProfileAvatarTypeEnum.im,
+      dateCreated: DateTime.now(),
+      name: 'Dummy User',
+      slug: 'dummy-user',
+      profileAvatarSeed: 'dummy-seed',
+      profileImage: 'https://placehold.co/400',
+      circleCount: 0,
     ),
-    category: '',
+    category: 'Dummy Category',
+    subscribers: 0,
+    recurring: 'Dummy Recurring',
+    price: 0,
+    nextEvents: [
+      NextEventSchema(
+        start: DateTime.now(),
+        link: 'https://placehold.co/400',
+        seatsLeft: 0,
+        slug: 'dummy-event',
+        title: 'Dummy Event',
+        attending: false,
+        calLink: 'https://placehold.co/400',
+        duration: 0,
+        meetingProvider: MeetingProviderEnum.googleMeet,
+        cancelled: false,
+        open: true,
+        joinable: true,
+      ),
+    ],
   );
 }
 
@@ -64,13 +111,23 @@ class SpaceCard extends StatelessWidget {
     );
   }
 
-  final SpaceDetailSchema space;
+  static Widget shimmer({bool compact = false}) {
+    return Shimmer.fromColors(
+      baseColor: Colors.grey.shade300,
+      highlightColor: Colors.grey.shade100,
+      child: SpaceCard(space: _dummySpaceDetail(), compact: compact),
+    );
+  }
+
+  final MobileSpaceDetailSchema space;
   final bool compact;
   final VoidCallback? onTap;
 
   @override
   Widget build(BuildContext context) {
     final theme = Theme.of(context);
+
+    final nextEvent = space.nextEvents.firstOrNull;
 
     return AspectRatio(
       aspectRatio: 1.38,
@@ -82,11 +139,11 @@ class SpaceCard extends StatelessWidget {
           onTap:
               onTap ??
               () async {
-                if (space.nextEvent != null) {
+                if (nextEvent != null) {
                   await context.push(
                     RouteNames.spaceEvent(
                       space.slug,
-                      space.nextEvent!.slug,
+                      nextEvent.slug,
                     ),
                   );
                 } else {
@@ -125,21 +182,23 @@ class SpaceCard extends StatelessWidget {
                     if (!isContentVisible) {
                       return const SizedBox.shrink();
                     }
-                    final seatsLeft = DefaultTextStyle.merge(
-                      style: TextStyle(
-                        color: Colors.white,
-                        fontSize: 10,
-                        fontWeight: FontWeight.bold,
-                        shadows: kElevationToShadow[4],
-                      ),
-                      child: buildSeatsLeftText(
-                        space.nextEvent?.seatsLeft ?? 0,
-                      ),
-                    );
+                    final seatsLeft = nextEvent != null
+                        ? DefaultTextStyle.merge(
+                            style: TextStyle(
+                              color: Colors.white,
+                              fontSize: 10,
+                              fontWeight: FontWeight.bold,
+                              shadows: kElevationToShadow[4],
+                            ),
+                            child: buildSeatsLeftText(
+                              nextEvent.seatsLeft,
+                            ),
+                          )
+                        : const SizedBox.shrink();
                     return Column(
                       crossAxisAlignment: CrossAxisAlignment.start,
                       children: [
-                        if (space.nextEvent != null)
+                        if (nextEvent != null)
                           Container(
                             padding: const EdgeInsetsDirectional.all(8),
                             decoration: BoxDecoration(
@@ -157,7 +216,7 @@ class SpaceCard extends StatelessWidget {
                                 ),
                                 Flexible(
                                   child: Text(
-                                    buildTimeLabel(space.nextEvent!.start),
+                                    buildTimeLabel(nextEvent.start),
                                     style: TextStyle(
                                       color: Colors.white,
                                       fontSize: 8,
@@ -183,9 +242,9 @@ class SpaceCard extends StatelessWidget {
                           maxLines: 2,
                           overflow: TextOverflow.fade,
                         ),
-                        if (space.nextEvent?.title != null)
+                        if (nextEvent?.title != null)
                           AutoSizeText(
-                            'Next: ${space.nextEvent!.title}',
+                            'Next: ${nextEvent!.title}',
                             style: TextStyle(
                               color: Colors.white,
                               fontSize: 9,
@@ -254,6 +313,7 @@ class SmallSpaceCard extends StatelessWidget {
     this.onTap,
     super.key,
   });
+
   factory SmallSpaceCard.fromEventDetailSchema(
     EventDetailSchema event, {
     VoidCallback? onTap,
@@ -264,12 +324,23 @@ class SmallSpaceCard extends StatelessWidget {
     );
   }
 
-  final SpaceDetailSchema space;
+  final MobileSpaceDetailSchema space;
   final VoidCallback? onTap;
+
+  static Widget shimmer() {
+    return Shimmer.fromColors(
+      baseColor: Colors.grey.shade300,
+      highlightColor: Colors.grey.shade100,
+      child: SmallSpaceCard(space: _dummySpaceDetail()),
+    );
+  }
 
   @override
   Widget build(BuildContext context) {
     final theme = Theme.of(context);
+
+    final nextEvent = space.nextEvents.firstOrNull;
+
     return InkWell(
       borderRadius: BorderRadius.circular(20),
       highlightColor: theme.colorScheme.secondary.withValues(alpha: 0.1),
@@ -311,13 +382,13 @@ class SmallSpaceCard extends StatelessWidget {
           ),
           PositionedDirectional(
             top: 10,
-            start: 10,
-            end: 10,
+            start: 12,
+            end: 12,
             bottom: 10,
             child: Column(
               crossAxisAlignment: CrossAxisAlignment.start,
               children: [
-                if (space.nextEvent != null)
+                if (nextEvent != null)
                   Container(
                     padding: const EdgeInsetsDirectional.all(8),
                     decoration: BoxDecoration(
@@ -325,7 +396,7 @@ class SmallSpaceCard extends StatelessWidget {
                       borderRadius: BorderRadius.circular(20),
                     ),
                     child: Text(
-                      buildTimeLabel(space.nextEvent!.start),
+                      buildTimeLabel(nextEvent.start),
                       style: TextStyle(
                         color: Colors.white,
                         fontSize: 8,
@@ -335,15 +406,17 @@ class SmallSpaceCard extends StatelessWidget {
                     ),
                   ),
                 const Spacer(),
-                Text(
-                  '${space.nextEvent?.seatsLeft ?? 'No'} seats left',
-                  style: TextStyle(
-                    color: Colors.white,
-                    fontSize: 14,
-                    shadows: kElevationToShadow[4],
+                if (nextEvent != null) ...[
+                  DefaultTextStyle.merge(
+                    style: TextStyle(
+                      color: Colors.white,
+                      fontSize: 12,
+                      shadows: kElevationToShadow[4],
+                    ),
+                    child: buildSeatsLeftText(nextEvent.seatsLeft),
                   ),
-                ),
-                const SizedBox(height: 5),
+                  const SizedBox(height: 5),
+                ],
                 AutoSizeText(
                   space.title,
                   style: TextStyle(
